Change Index.js files to Page.js. Create Tasks nested directory and move Page.ja & Form.js.
