+++ client/src/components/Sales/Page.js
