+++ server/src/models/Task.js
@@ -5,7 +5,7 @@ let taskSchema = new mongoose.Schema({
   _creator: { type: mongoose.Schema.Types.ObjectId, ref: "Project" },
   name: { type: String, required: true },
   hours: { type: Number, min: 1, max: 8, required: true },
-  payment_type: { type: String, required: true },
+  paymentType: { type: String, required: true },
   price: { type: Number, required: true },
   vat: { type: Number, min: 1, max: 100, required: true },
 
@@ -18,7 +18,7 @@ taskSchema.post('save', function(doc, next) {
   Project.findByIdAndUpdate(this._creator, { $push: { tasks: this._id} }, { new: true }, function(err, project) {
     if (err) {
       errors: {
-        cant_update_project: {
+        cantUpdateProject: {
           message: err
         } 
       }
